19-05-2022 10h24

+++ 02_ft_printf/srcs/ft_puthex.c
@@ -6,7 +6,7 @@
 /*   By: halvarez <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/10 11:09:32 by halvarez          #+#    #+#             */
-/*   Updated: 2022/05/18 18:00:16 by halvarez         ###   ########.fr       */
+/*   Updated: 2022/05/19 10:05:45 by halvarez         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,7 +35,7 @@ void	ft_ptr_to_hex(va_list argp, int *count)
 void	ft_long_to_hex(va_list argp, int *count, char conv)
 {
 	char			*hex;
-	long	n;
+	unsigned int	n;
 
 	n = va_arg(argp, long);
 	hex = NULL;
@@ -43,7 +43,7 @@ void	ft_long_to_hex(va_list argp, int *count, char conv)
 		hex = "0123456789abcdef";
 	else if (conv == 'X')
 		hex = "0123456789ABCDEF";
-	ft_puthex((unsigned int)n, hex, count);
+	ft_puthex(n, hex, count);
 }
 
 void	ft_puthex(unsigned long n, char *base, int *count)
